+++ packages/dredd/lib/reporters/ApiaryReporter.js
@@ -1,8 +1,8 @@
 import clone from 'clone';
 import { v4 as generateUuid } from 'uuid';
 import os from 'os';
-import request from 'request';
 
+import request from '../httpClient';
 import logger from '../logger';
 import reporterOutputLogger from './reporterOutputLogger';
 import packageData from '../../package.json';